HAPP-1332: Create FetchCommentsUseCase
